Gorgon Imaging
- Removed all references to SharpDX from imaging API.
- Separated out image codec plug in support into its own assembly to allow AOT for base image API.
- Separated out image <-> GDI functionality into separate assembly in case it is not needed.
- Reorganized structure of graphics APIs on file system.
- Updated many functions to use GorgonPtr instead of unsafe pointers where appropriate and applicable.
- Moved BufferFormat, GorgonFormatInfo and GorgonPitchLayout to imaging API. This makes more sense as these values are related to images exclusively.
- Updated 16bpp 4 bit per channel conversion functions to use 256 bit SIMD if available. Provides a substantial performance increase.
- Added new types, GorgonPixels and GorgonPixel.  Pixels is a container for common predefined pixel formats, and the GorgonPixel type is a base class for pixel formats. This object is also used to read and write pixels from and to a IGorgonImageBuffer. These can be extended to provide more functionality.
- Updated IGorgonImage SizeInBytes and any other total image size related functionality to be a Int64 value for large images.
- Dropped dependency on BCnEncoder.NET. No plans yet to allow encoding of BC compressed images (external tools do a much better job).

Gorgon.Core

- Removed any use of ReadOnly/Span<T> types as they are limited to 2 GB in size and use GorgonPtr exclusively instead (pointers can still use spans).
- Added a stream object for native memory.
- Added method/operators to convert the type parameter in GorgonPtr<T> to another type (e.g. (GorgonPtr<byte>)gorgonPtrInt32;)
- Removed ToBytePointer, this is now supplanted by ToPointerType.

And as always... general cleanup.

+++ Gorgon/Gorgon.Renderers/Gorgon2D/_Internal/BatchRenderer.cs
@@ -136,7 +136,7 @@ internal sealed class BatchRenderer
                                                                                      });
 
         IndexBuffer = new GorgonIndexBuffer(Graphics,
-                                            new GorgonIndexBufferInfo(indices.Length)
+                                            new GorgonIndexBufferInfo((int)indices.Length)
                                             {
                                                 Usage = ResourceUsage.Immutable,
                                                 Binding = VertexIndexBufferBinding.None
